fix: revert content-disposition parse logic and make fallback naturally
